+++ src/extension.ts
@@ -3,6 +3,8 @@ import * as vscode from 'vscode';
 import { CleanAircraftCfg } from './tools/aircraft-cfg/clean-v2';
 import { CleanFlightplan } from './tools/flightplan/clean';
 import { ChangeAircraftNumber } from './tools/flightplan/change-ac-number';
+import { RenumberAddOnsCfg } from './tools/add-ons-cfg/renumber';
+import { RenumberSceneryCfg } from './tools/scenery-cfg/renumber';
 
 export function activate(context: vscode.ExtensionContext) {
 	// TODO limit commands to FS files
@@ -27,6 +29,18 @@ export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(
 		vscode.commands.registerCommand('extension.changeAircraftNumber', async () => {
 			ChangeAircraftNumber();
+		})
+	);
+
+	context.subscriptions.push(
+		vscode.commands.registerCommand('extension.renumberAddOnsCfg', async () => {
+			RenumberAddOnsCfg();
+		})
+	);
+
+	context.subscriptions.push(
+		vscode.commands.registerCommand('extension.renumberSceneryCfg', async () => {
+			RenumberSceneryCfg();
 		})
 	);
 }