+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+### 1.0.6
+
+- `Renumber scenery.cfg` added. Renumbers the `[Area.xxx]` entries sequentially and sets the `Layer=x` properties to the same number.
+- `Renumber add-ons.cfg` added. Renumbers the `[Package.x]` entries sequentially.
+
 ### 1.0.5
 
 - `Clean Aircraft.cfg` rewritten completely, can now take care of empty lines and sort entry lines.